DJ: Renamed packages and directories, and amended build.gradle file
