+++ DietManagerLogger.log
@@ -2,25 +2,25 @@
 <!DOCTYPE log SYSTEM "logger.dtd">
 <log>
 <record>
-  <date>2020-03-16T16:26:31.455603Z</date>
-  <millis>1584375991455</millis>
-  <nanos>603000</nanos>
+  <date>2020-03-18T08:49:46.462279100Z</date>
+  <millis>1584521386462</millis>
+  <nanos>279100</nanos>
   <sequence>0</sequence>
-  <logger>seedu.duke.InfoLogger</logger>
+  <logger>seedu.dietmanager.InfoLogger</logger>
   <level>INFO</level>
-  <class>seedu.duke.InfoLogger</class>
+  <class>seedu.dietmanager.InfoLogger</class>
   <method>logExecuteProgramme</method>
   <thread>1</thread>
   <message>Starting Diet Manager</message>
 </record>
 <record>
-  <date>2020-03-16T16:27:01.244892700Z</date>
-  <millis>1584376021244</millis>
-  <nanos>892700</nanos>
+  <date>2020-03-18T08:50:22.273014800Z</date>
+  <millis>1584521422273</millis>
+  <nanos>14800</nanos>
   <sequence>1</sequence>
-  <logger>seedu.duke.InfoLogger</logger>
+  <logger>seedu.dietmanager.InfoLogger</logger>
   <level>INFO</level>
-  <class>seedu.duke.InfoLogger</class>
+  <class>seedu.dietmanager.InfoLogger</class>
   <method>logExitProgramme</method>
   <thread>1</thread>
   <message>Exiting Diet Manager</message>
