+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'com.github.johnrengelman.shadow' version '5.1.0'
 }
 
-group 'seedu.duke'
+group 'seedu.dietmanager'
 version '0.1.0'
 
 repositories {
@@ -21,7 +21,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Duke"
+    mainClassName = "seedu.dietmanager.DietManager"
 }
 
 shadowJar {
@@ -35,6 +35,7 @@ checkstyle {
     toolVersion = '8.23'
 }
 
-run{
+run {
     standardInput = System.in
+    enableAssertions = true
 }
