+++ src/main/java/seedu/dietmanager/DailyFoodRecord.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.dietmanager;
 
 import java.util.ArrayList;
 import java.time.LocalDate;
