+++ src/main/java/seedu/dietmanager/Food.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.dietmanager;
 
 public class Food {
     String foodName;
