+++ src/main/java/seedu/dietmanager/commands/Command.java
@@ -1,7 +1,7 @@
-package seedu.duke.commands;
+package seedu.dietmanager.commands;
 
-import seedu.duke.Profile;
-import seedu.duke.ui.UI;
+import seedu.dietmanager.Profile;
+import seedu.dietmanager.ui.UI;
 
 /**
  * Command is a public abstract class providing a skeletal implementation for
