+++ src/main/java/seedu/dietmanager/commands/ProfileCommand.java
@@ -1,7 +1,7 @@
-package seedu.duke.commands;
+package seedu.dietmanager.commands;
 
-import seedu.duke.Profile;
-import seedu.duke.ui.UI;
+import seedu.dietmanager.Profile;
+import seedu.dietmanager.ui.UI;
 
 public class ProfileCommand extends Command {
