+++ src/main/java/seedu/dietmanager/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.dietmanager.exceptions;
 
 public class InvalidCommandException extends Exception {
 }
