+++ src/main/java/seedu/dietmanager/exceptions/InvalidFormatException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exceptions;
+package seedu.dietmanager.exceptions;
 
 public class InvalidFormatException extends Exception {
 }
