+++ src/main/java/seedu/dietmanager/ui/UI.java
@@ -1,4 +1,4 @@
-package seedu.duke.ui;
+package seedu.dietmanager.ui;
 
 import java.util.Scanner;
