+++ src/test/java/seedu/dietmanager/DietManagerTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.dietmanager;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class DietManagerTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
